Return Firestore types to API callers
Originally, the intent was to abstract away most of the Firestore
internals (e.g. DocumentSnapshot, etc) so that callers wouldn't have to
manually deal with transforming them to the desired end types. An
additional goal was to give clients the ability to use Firestore without
having to directly depend on the Firestore SDK.

However, it turns out that, for several reasons, clients would still
need to depend on the Firestore SDK — for instance, for creating filters
or for accessing the DocumentSnapshot to apply transformations to the
data returned from the backend. Therefore, it made less and less sense
to try to hide these internals from users of this library.

On top of that, returning DocumentSnapshot (and DocumentChange) directly
in flows vs the finalized data type that the client desires allows us to
make these regular functions instead of inline functions, which were
previously necessary due to the reification of the type parameter T.